Задание Отрисуй меня полностью
