+++ js/main.js
@@ -1,4 +1,8 @@
-import { generatePhotosData } from './photos.js';
+import { generatePhotosData } from './data.js';
 
 const photosData = generatePhotosData();
 console.log(photosData);
+
+import { renderThumbnails } from './thumbnails.js';
+
+renderThumbnails();
