--- js/photos.js
@@ -1,21 +1,0 @@
-import { descriptions } from './data.js';
-import { getRandomInt, getRandomArrayElement } from './utils.js';
-import { generateComments } from './comments.js';
-
-export function generatePhotosData() {
-  const photos = [];
-
-  for (let i = 1; i <= 25; i++) {
-    const photo = {
-      id: i,
-      url: `photos/${i}.jpg`,
-      description: getRandomArrayElement(descriptions),
-      likes: getRandomInt(15, 200),
-      comments: generateComments(getRandomInt(0, 5))
-    };
-
-    photos.push(photo);
-  }
-
-  return photos;
-}
